+++ src/ArduinoJson/Polyfills/type_traits/is_floating_point.hpp
@@ -5,15 +5,16 @@
 #pragma once
 
 #include "integral_constant.hpp"
+#include "is_same.hpp"
+#include "remove_cv.hpp"
 
 namespace ARDUINOJSON_NAMESPACE {
 
-template <typename>
-struct is_floating_point : false_type {};
+template <class T>
+struct is_floating_point
+    : integral_constant<
+          bool,  //
+          is_same<float, typename remove_cv<T>::type>::value ||
+              is_same<double, typename remove_cv<T>::type>::value> {};
 
-template <>
-struct is_floating_point<float> : true_type {};
-
-template <>
-struct is_floating_point<double> : true_type {};
 }  // namespace ARDUINOJSON_NAMESPACE