+++ src/ArduinoJson/Polyfills/type_traits/is_unsigned.hpp
@@ -5,33 +5,24 @@
 #pragma once
 
 #include "integral_constant.hpp"
+#include "is_same.hpp"
+#include "remove_cv.hpp"
+
 namespace ARDUINOJSON_NAMESPACE {
 
-template <typename>
-struct is_unsigned : false_type {};
-
-template <>
-struct is_unsigned<bool> : true_type {};
-
-template <>
-struct is_unsigned<unsigned char> : true_type {};
-
-template <>
-struct is_unsigned<unsigned short> : true_type {};
-
-template <>
-struct is_unsigned<unsigned int> : true_type {};
-
-template <>
-struct is_unsigned<unsigned long> : true_type {};
-
+// clang-format off
+template <typename T>
+struct is_unsigned : integral_constant<bool,
+    is_same<typename remove_cv<T>::type, unsigned char>::value ||
+    is_same<typename remove_cv<T>::type, unsigned short>::value ||
+    is_same<typename remove_cv<T>::type, unsigned int>::value ||
+    is_same<typename remove_cv<T>::type, unsigned long>::value ||
 #if ARDUINOJSON_HAS_INT64
-template <>
-struct is_unsigned<unsigned __int64> : true_type {};
+    is_same<typename remove_cv<T>::type, unsigned __int64>::value ||
 #endif
-
 #if ARDUINOJSON_HAS_LONG_LONG
-template <>
-struct is_unsigned<unsigned long long> : true_type {};
+    is_same<typename remove_cv<T>::type, unsigned long long>::value ||
 #endif
+    is_same<typename remove_cv<T>::type, bool>::value> {};
+// clang-format on
 }  // namespace ARDUINOJSON_NAMESPACE